pyproject.toml
[project]
name = "lab-manager"
version = "0.1.5"
description = "Lab inventory management with OCR document intake"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.115.0",
  "sqlmodel>=0.0.22",
  "sqlalchemy[asyncio]>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.14.0",
  "sqladmin>=0.20.0",
  "pydantic-settings>=2.6.0",
  "meilisearch>=0.33.0",
  "instructor>=1.7.0",
  "google-genai>=1.0.0",
  "pillow>=11.2.1",
  "python-multipart>=0.0.18",
  "python-ulid>=3.0.0",
  "bcrypt>=4.0.0",
  "itsdangerous>=2.2.0",
  "structlog>=24.0.0",
  "slowapi>=0.1.9",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=0.24.0",
  "pytest-bdd>=8.0.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.28.0",
  "ruff>=0.11.0",
  "mypy>=1.10.0",
  "bandit>=1.8.0",
  "testcontainers[postgres]>=4.0.0",
  "pip-audit>=2.7.0",
]

[tool.uv]
constraint-dependencies = ["pyasn1>=0.6.3"]

[tool.ruff]
line-length = 120
extend-exclude = ["scripts/", "ocr-benchmark/"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4"]
ignore = ["SIM102", "SIM105", "SIM108", "B007", "B008", "B017", "B904", "B905", "C401", "C408", "C416", "UP042", "UP047"]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
warn_return_any = false
warn_unused_ignores = false
warn_redundant_casts = false
# Disable noisy error codes from SQLAlchemy metaprogramming and untyped code.
# These will be re-enabled incrementally as the codebase is annotated.
disable_error_code = [
  "arg-type",
  "attr-defined",
  "assignment",
  "call-overload",
  "dict-item",
  "list-item",
  "operator",
  "union-attr",
  "return-value",
  "var-annotated",
]
